Move resetReceived*Requests to HermesMock
