+++ hermes-mock/src/main/java/pl/allegro/tech/hermes/mock/HermesMock.java
@@ -2,20 +2,36 @@ package pl.allegro.tech.hermes.mock;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.tomakehurst.wiremock.WireMockServer;
+import com.github.tomakehurst.wiremock.http.Request;
+import com.github.tomakehurst.wiremock.matching.RequestPattern;
+import com.github.tomakehurst.wiremock.matching.RequestPatternBuilder;
+import com.github.tomakehurst.wiremock.matching.ValueMatcher;
+import org.apache.avro.Schema;
+import pl.allegro.tech.hermes.mock.matching.ContentMatchers;
 
+import java.util.function.Predicate;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
+import static com.github.tomakehurst.wiremock.http.RequestMethod.POST;
+import static pl.allegro.tech.hermes.mock.HermesMockHelper.startsWith;
 
 public class HermesMock {
+
+    private static final String APPLICATION_JSON = "application/json";
+    private static final String AVRO_BINARY = "avro/binary";
+
     private final WireMockServer wireMockServer;
 
     private final HermesMockDefine hermesMockDefine;
     private final HermesMockExpect hermesMockExpect;
     private final HermesMockQuery hermesMockQuery;
+    private final HermesMockHelper hermesMockHelper;
 
     private HermesMock(int port, int awaitSeconds, ObjectMapper objectMapper) {
         wireMockServer = new WireMockServer(port);
 
-        HermesMockHelper hermesMockHelper = new HermesMockHelper(wireMockServer, objectMapper);
+        hermesMockHelper = new HermesMockHelper(wireMockServer, objectMapper);
         hermesMockDefine = new HermesMockDefine(hermesMockHelper);
         hermesMockExpect = new HermesMockExpect(hermesMockHelper, awaitSeconds);
         hermesMockQuery = new HermesMockQuery(hermesMockHelper);
@@ -35,6 +51,25 @@ public class HermesMock {
 
     public void resetReceivedRequest() {
         wireMockServer.resetRequests();
+    }
+
+    public <T> void resetReceivedAvroRequests(String topicName, Schema schema, Class<T> clazz, Predicate<T> predicate) {
+        ValueMatcher<Request> valueMatcher = ContentMatchers.matchAvro(hermesMockHelper, predicate, schema, clazz);
+        resetReceivedRequests(topicName, AVRO_BINARY, valueMatcher);
+    }
+
+    public <T> void resetReceivedJsonRequests(String topicName, Class<T> clazz, Predicate<T> predicate) {
+        ValueMatcher<com.github.tomakehurst.wiremock.http.Request> valueMatcher = ContentMatchers.matchJson(hermesMockHelper, predicate, clazz);
+        resetReceivedRequests(topicName, APPLICATION_JSON, valueMatcher);
+    }
+
+    private void resetReceivedRequests(String topicName, String contentType, ValueMatcher<com.github.tomakehurst.wiremock.http.Request> valueMatcher) {
+        RequestPattern requestPattern = RequestPatternBuilder
+                .newRequestPattern(POST, urlEqualTo("/topics/" + topicName))
+                .withHeader("Content-Type", startsWith(contentType))
+                .andMatching(valueMatcher)
+                .build();
+        wireMockServer.removeServeEventsMatching(requestPattern);
     }
 
     public void resetMappings() {
