+++ hermes-mock/src/test/groovy/pl/allegro/tech/hermes/mock/HermesMockAvroTest.groovy
@@ -288,33 +288,6 @@ class HermesMockAvroTest extends Specification {
             1 == hermes.query().countMatchingAvroMessages(topicName, schema, TestMessage, filter)
     }
 
-    def "should reset received requests with Avro messages"() {
-        given:
-        def topicName = "my-test-avro-topic"
-        hermes.define().avroTopic(topicName)
-        publish(topicName, new TestMessage("test-key", "test-value"))
-        assert hermes.query().countAvroMessages(topicName) == 1
-
-        when:
-        hermes.query().resetReceivedAvroRequests(topicName, schema, TestMessage, {it -> it.key == "test-key"})
-
-        then:
-        hermes.query().countAvroMessages(topicName) == 0
-    }
-
-    def "should not reset received requests with Avro messages when predicate does not match"() {
-        given:
-        def topicName = "my-test-avro-topic"
-        hermes.define().avroTopic(topicName)
-        publish(topicName, new TestMessage("test-key", "test-value"))
-
-        when:
-        hermes.query().resetReceivedAvroRequests(topicName, schema, TestMessage, {it -> it.key == "different-test-key"})
-
-        then:
-        hermes.query().countAvroMessages(topicName) == 1
-    }
-
     def asAvro(TestMessage message) {
         return jsonAvroConverter.convertToAvro(message.asJson().bytes, schema)
     }
